+++ src/core/resources/tilemap.c
@@ -23,6 +23,8 @@ TilesetPack* CreateTilesetsPack(const char* name, const char* filename)
 	pack->name = name;
 	pack->nextTilesetIndex = 1;
 	pack->tilesetsCount = 0;
+	pack->tileCount = 0;
+	pack->tiles = NULL;
 	pack->tilesets = NULL;
 
 	unsigned int hasError = 0; // indica que no existen errores en la carga
@@ -53,6 +55,7 @@ TilesetPack* CreateTilesetsPack(const char* name, const char* filename)
 					if (strcmp(key, "ts") == 0)
 					{
 						Tileset* newTileset = CreateTileset(val, pack->nextTilesetIndex);
+						printf("TEXTURA ID AL SER CARGADA: %d\n", newTileset->texture.id);
 						if (newTileset != NULL)
 						{
 							Tileset** tsMemTemp = (Tileset**)realloc(pack->tilesets, (size_t)(pack->tilesetsCount + 1) * sizeof(Tileset*));
@@ -85,7 +88,6 @@ TilesetPack* CreateTilesetsPack(const char* name, const char* filename)
 	if (hasError > 0)
 	{
 		printf("Se han encontrado errores en la carga, imposible generar el paquete de conjuntos de mosaicos.\n");
-
 		if(pack->tilesetsCount > 0)
 		{
 			for (int i = 0; i < pack->tilesetsCount; i++)
@@ -106,13 +108,60 @@ TilesetPack* CreateTilesetsPack(const char* name, const char* filename)
 		return NULL;
 	}
 
-	// generar los tilests
+	// generar los tiles
+	unsigned int totalTiles = 0;
+	for (int i = 0; i < pack->tilesetsCount; i++)
+		totalTiles += pack->tilesets[i]->tilesCount;
+	printf("TOTAL TILES %d\n", totalTiles);
+
+	Tile** tilesMemTemp = (Tile**)realloc(pack->tiles, (size_t)(totalTiles + 1) * sizeof(Tile*));
+	if (tilesMemTemp == NULL)
+	{
+		for (int i = 0; i < pack->tilesetsCount; i++)
+		{
+			if (pack->tilesets[i] != NULL)
+				free(pack->tilesets[i]);
+		}
+		free(pack);
+		return NULL;
+	}
+	pack->tiles = tilesMemTemp;
+
+	pack->tiles[pack->tileCount] = NULL;
+	pack->tileCount += 1;
+	for (int i = 0; i < pack->tilesetsCount; i++)
+	{
+		unsigned int rows = pack->tilesets[i]->texture.height / pack->tilesets[i]->tileHeight;
+		unsigned int columns = pack->tilesets[i]->texture.width / pack->tilesets[i]->tileWidth;
+		printf("ROWS: %d\n", rows);
+		printf("COLUMNS: %d\n", columns);
+
+		for (int ty = 0; ty < rows; ty++)
+		{
+			for (int tx = 0; tx < columns; tx++)
+			{
+				Tile* newTile = (Tile*)malloc(sizeof(Tile));
+				if (newTile != NULL)
+				{
+					newTile->x = tx * pack->tilesets[i]->tileWidth;
+					newTile->y = ty * pack->tilesets[i]->tileHeight;
+					newTile->width = pack->tilesets[i]->tileWidth;
+					newTile->height = pack->tilesets[i]->tileHeight;
+					pack->tiles[pack->tileCount] = newTile;
+					printf("TILES AGREGADO CON: %d, %d, %d, %d\n", newTile->x, newTile->y, newTile->width, newTile->height);
+				}
+				pack->tileCount += 1;
+			}
+		}
+	}
 
 	return pack;
 }
 
 int InsertTilesetToPack(TilesetPack* pack, const char* filename)
 {
+	if (pack == NULL) return 0;
+
 	Tileset* tileset = CreateTileset(filename, pack->nextTilesetIndex);
 	if (tileset == NULL) return 0;
 
@@ -133,10 +182,19 @@ Tileset* GetTileset(TilesetPack* pack, unsigned int index)
 
 	for (int i = 0; i < pack->tilesetsCount; i++)
 	{
-		if (index >= pack->tilesets[i]->initialIndex && index <= pack->tilesets[i]->finalIndex)
-			return pack->tilesets[i];
+		printf("DEBUG TILESET 0\n");
+		printf("DEBUG TILESET COUNT %d\n", pack->tilesetsCount);
+		if(pack->tilesets[i] != NULL)
+		{
+			if (index >= pack->tilesets[i]->initialIndex && index <= pack->tilesets[i]->finalIndex)
+			{
+				printf("DEBUG TILESET 1\n");
+				return pack->tilesets[i];
+			}
+		}
+		printf("DEBUG TILESET 2\n");
 	}
-
+	printf("DEBUG TILESET 4\n");
 	return NULL;
 }
 
@@ -234,6 +292,7 @@ Tileset* CreateTileset(const char* filename, unsigned int initialIndex)
 	mxml_node_t* textureNode = mxmlFindElement(tsNode, root, "image", NULL, NULL, MXML_DESCEND_ALL);
 	const char* textureFilename = mxmlElementGetAttr(textureNode, "source");
 	tileset->texture = LoadTexture(textureFilename);
+	printf("TILESET TEXTURE: %d\n", tileset->texture.id);
 
 	// descargar datos de memoria
 	mxmlOptionsDelete(options);
@@ -250,10 +309,6 @@ TileMap* CreateMap(const char* name, const char* filename, TilesetPack* pack)
 		printf("Error al asignar memoria para el tilemap: %s\n", name);
 		return NULL;
 	}
-
-	tmap->name = name;
-	tmap->layers = NULL;
-	tmap->layerCount = 0;
 
 	// cargar el archivo del mapa
 	FILE* xmlFile = fopen(filename, "r");
@@ -289,28 +344,36 @@ TileMap* CreateMap(const char* name, const char* filename, TilesetPack* pack)
 	}
 
 	// Datos del mapa
+	tmap->name = name;
+	tmap->layers = NULL;
+	tmap->layerCount = 0;
 	tmap->tileWidth = atoi(mxmlElementGetAttr(mapNode, "tilewidth"));
 	tmap->tileHeight = atoi(mxmlElementGetAttr(mapNode, "tileheight"));
 	tmap->mapWidth = atoi(mxmlElementGetAttr(mapNode, "width"));
 	tmap->mapHeight = atoi(mxmlElementGetAttr(mapNode, "height"));
+	printf("DEBUG 0\n");
 
 	// cargar las capas
 	unsigned int hasErrors = 0;
 	mxml_node_t* layerNode = mxmlFindElement(tree, tree, "layer", NULL, NULL, MXML_DESCEND_ALL);
+	printf("DEBUG 1\n");
 	while (layerNode != NULL)
 	{
 		mxml_node_t* layerData = mxmlFindElement(layerNode, layerNode, "data", NULL, NULL, MXML_DESCEND_ALL);
 		if (layerData == NULL) hasErrors++;
+		printf("DEBUG 2\n");
 
 		TiledLayer* layer = CreateLayer(layerData, tmap->mapWidth, tmap->mapHeight);
 		if (layer == NULL) hasErrors++;
+		printf("DEBUG 3\n");
 
-		TiledLayer** layersMemTemp = (TiledLayer**)malloc((size_t)(tmap->layerCount + 1) * sizeof(TiledLayer*));
+		TiledLayer** layersMemTemp = (TiledLayer**)realloc(tmap->layers, (size_t)(tmap->layerCount + 1) * sizeof(TiledLayer*));
+		printf("DEBUG 4\n");
 		if (layersMemTemp != NULL && layer != NULL) 
 		{
 			tmap->layers = layersMemTemp;
 			tmap->layers[tmap->layerCount] = layer;
-			tmap->layerCount += 1; // hacer a visual studio feliz	
+			tmap->layerCount += 1; // hacer a visual studio feliz
 		}
 		else
 			hasErrors++;
@@ -318,6 +381,7 @@ TileMap* CreateMap(const char* name, const char* filename, TilesetPack* pack)
 
 		layerNode = mxmlFindElement(layerNode, tree, "layer", NULL, NULL, MXML_DESCEND_ALL);
 	}
+	printf("DEBUG 4\n");
 
 	if (hasErrors > 0)
 	{
@@ -335,6 +399,8 @@ TileMap* CreateMap(const char* name, const char* filename, TilesetPack* pack)
 		mxmlDelete(tree);
 		return NULL;
 	}
+	printf("DEBUG 5\n");
+
 	if (tmap->layerCount == 0)
 	{
 		printf("No se han encontrado capas para el mapa {%s}, imposible crear un mapa sin capas.\n", tmap->name);
@@ -351,46 +417,58 @@ TileMap* CreateMap(const char* name, const char* filename, TilesetPack* pack)
 		mxmlDelete(tree);
 		return NULL;
 	}
+	printf("DEBUG 6\n");
 
 	// crear textura
 	tmap->mapTexture = LoadRenderTexture(
 		tmap->tileWidth * tmap->mapWidth,
 		tmap->tileHeight * tmap->mapHeight
 	);
-
+	printf("DEBUG 8\n");
 	// Dibujar las capas
 	BeginTextureMode(tmap->mapTexture);
+	
 	ClearBackground(BLANK);
 
 	for (int i = 0; i < tmap->layerCount; i++)
 	{
+		printf("DEBUG 9\n");
 		unsigned int tileIndex = 0;
 		for (int ty = 0; ty < tmap->mapHeight; ty++)
 		{
+			printf("DEBUG 10\n");
 			for (int tx = 0; tx < tmap->mapWidth; tx++)
 			{
 				// Cero indica que el espacio está vacío.
+				printf("DEBUG 11\n");
 				if (tmap->layers[i]->data[tileIndex] != 0)
 				{
+					printf("DEBUG 12\n");
 					Tileset* tileset = GetTileset(pack, tmap->layers[i]->data[tileIndex]);
+					printf("DEBUG 13\n");
 					if (tileset != NULL)
 					{
+						printf("DEBUG 14\n");
 						if (pack->tiles[tmap->layers[i]->data[tileIndex]] != NULL)
 						{
+							printf("DEBUG 15\n");
+							printf("UN DATO DEL PACK: %d\n", pack->tiles[0]->width);
 							Rectangle src = {
 								pack->tiles[tmap->layers[i]->data[tileIndex]]->x,
 								pack->tiles[tmap->layers[i]->data[tileIndex]]->y,
 								tmap->tileWidth,
 								tmap->tileHeight
 							};
+							printf("DEBUG 16\n");
 							Rectangle dst = {
 								tx * tmap->tileWidth,
 								ty * tmap->tileHeight,
 								tmap->tileWidth,
 								tmap->tileHeight
 							};
-
+							printf("DEBUG 17\n");
 							DrawTexturePro(tileset->texture, src, dst, (Vector2) { 0.0f, 0.0f }, 0.0f, WHITE);
+							printf("DEBUG 18\n");
 						}
 						tileIndex++;
 					}
@@ -406,6 +484,9 @@ TileMap* CreateMap(const char* name, const char* filename, TilesetPack* pack)
 	return tmap;
 }
 
+/*
+EL ERROR ESTÁ EN QUE EL VALOR DE TILE_WIDTH Y TILE_HEIGHT DE TILESET ESTÁ EN VALOR 0.
+*/
 TiledLayer* CreateLayer(mxml_node_t* data, unsigned int mapWidth, unsigned int mapHeight)
 {
 	if (data == NULL) return NULL;
@@ -439,6 +520,7 @@ TiledLayer* CreateLayer(mxml_node_t* data, unsigned int mapWidth, unsigned int m
 	}
 
 	free(dataCopy);
+	printf("Capa creada\n");
 	return layer;
 }
 
@@ -516,6 +598,17 @@ void UnloadTilesetsPack(TilesetPack* pack)
 
 	// eliminar el paquete
 	free(pack);
+}
+
+void UnloadTreeTilesetsPackNode(TilesetPackNode* node)
+{
+	if (node == NULL) return;
+
+	UnloadTreeTilesetsPackNode(node->left);
+	UnloadTreeTilesetsPackNode(node->right);
+
+	UnloadTilesetsPack(node->pack);
+	free(node);
 }
 
 void UnloadTreeTileMapNode(TileMapNode* node)