+++ src/main.c
@@ -29,7 +29,8 @@ int main()
 	Game game;
 	InitGame(&game, config);
 
-	AddLevel(&game, "test", 0, 0, 0, TL_Run);
+	//AddLevel(&game, "test", 0, 0, 0, TL_Run);
+	AddLevel(&game, "menu", 0, 0, 0, Menu_Run);
 
 	RunGame(&game);
 	FreeGame(&game);